+++ packages/core/src/types.ts
@@ -4,37 +4,37 @@ export type MaybeArray<T> = T | T[]
 
 export type EventFn<D, T> = (data: D, payload?: any, result?: any) => T
 
-export type EventFnConfig<T, K> = Extract<keyof T, string> | K
+export type EventFnDesign<T, K> = Extract<keyof T, string> | K
 
 // Action
 
 export type Action<D> = EventFn<D, any>
 
-export type ActionConfig<D, T> = EventFnConfig<T, Action<D>>
+export type ActionDesign<D, T> = EventFnDesign<T, Action<D>>
 
 // Condition
 
 export type Condition<D> = EventFn<D, boolean>
 
-export type ConditionConfig<D, T> = EventFnConfig<T, Condition<D>>
+export type ConditionDesign<D, T> = EventFnDesign<T, Condition<D>>
 
 // Result
 
 export type Result<D> = EventFn<D, any>
 
-export type ResultConfig<D, T> = EventFnConfig<T, Result<D>>
+export type ResultDesign<D, T> = EventFnDesign<T, Result<D>>
 
 // Async
 
 export type Async<D> = EventFn<D, Promise<any>>
 
-export type AsyncConfig<D, T> = EventFnConfig<T, Async<D>>
+export type AsyncDesign<D, T> = EventFnDesign<T, Async<D>>
 
 // Time
 
 export type Time<D> = EventFn<D, number>
 
-export type TimeConfig<D, T> = number | EventFnConfig<T, Time<D>>
+export type TimeDesign<D, T> = number | EventFnDesign<T, Time<D>>
 
 // Value
 
@@ -48,7 +48,7 @@ export type Values<D, V extends Record<string, Value<D>>> = {
 
 export type Target<D> = EventFn<D, string>
 
-export type TargetConfig<D> = string | EventFn<D, string>
+export type TargetDesign<D> = string | EventFn<D, string>
 
 // Send
 
@@ -56,13 +56,13 @@ export type Event = { event: string; payload: any }
 
 export type Send<D> = EventFn<D, Event>
 
-export type SendConfig<D> = string | Event | EventFn<D, Event>
+export type SendDesign<D> = string | Event | EventFn<D, Event>
 
 // Break
 
 export type Break<D> = EventFn<D, boolean>
 
-export type BreakConfig<D> = boolean | EventFn<D, boolean>
+export type BreakDesign<D> = boolean | EventFn<D, boolean>
 
 // Event Handler Object
 
@@ -83,59 +83,59 @@ export type EventHandlerObject<D> = {
 /**
  * What to do when an event is triggered. You can define an event as an array of these objects.
  */
-export type EventHandlerObjectConfig<D, R, C, A, T> = {
+export type EventHandlerObjectDesign<D, R, C, A, T> = {
   /**
    * The result(s) to compute before running conditions or actions. The returned value will be available as the third argument to event handler functions, such as conditions, actions, and further restults.
    */
-  get?: MaybeArray<ResultConfig<D, R>>
+  get?: MaybeArray<ResultDesign<D, R>>
   /**
    * The condition(s) that must all be true in order for this event handler object to run.
    */
-  if?: MaybeArray<ConditionConfig<D, C>>
+  if?: MaybeArray<ConditionDesign<D, C>>
   /**
    * The condition(s) that must be all be false in order for this event handler object to run.
    */
-  unless?: MaybeArray<ConditionConfig<D, C>>
+  unless?: MaybeArray<ConditionDesign<D, C>>
   /**
    * The condition(s) of which at least one must be true in order for this event handler object to run.
    */
-  ifAny?: MaybeArray<ConditionConfig<D, C>>
+  ifAny?: MaybeArray<ConditionDesign<D, C>>
   /**
    * The action(s) to perform. These actions can mutate the data property. Note: Defining actions will cause this event to trigger an update.
    */
-  do?: MaybeArray<ActionConfig<D, A>>
+  do?: MaybeArray<ActionDesign<D, A>>
   /**
    * The "secret" action(s) to perform. These actions cannot mutate the data property. Note: Defining "secret" actions will NOT cause this event to trigger an update.
    */
-  secretlyDo?: MaybeArray<ActionConfig<D, A>>
+  secretlyDo?: MaybeArray<ActionDesign<D, A>>
   /**
    * A transition target: either a state's name or path. Defining a target will cause this event to trigger an update.
    */
-  to?: TargetConfig<D>
+  to?: TargetDesign<D>
   /**
    * An event name and (optionally) payload to send to the state.
    */
-  send?: SendConfig<D>
+  send?: SendDesign<D>
   /**
    * A delay (in seconds) to wait before running this handler object.
    */
-  wait?: TimeConfig<D, T>
+  wait?: TimeDesign<D, T>
   /**
    * Whether this item should stop the event's other handlers objects from running.
    */
-  break?: BreakConfig<D>
+  break?: BreakDesign<D>
   /**
    * An event handler to run instead if this event handler object did not pass its conditions.
    */
-  else?: EventHandlerConfig<D, R, C, A, T>
+  else?: EventHandlerDesign<D, R, C, A, T>
 }
 
 // Event Handler
 
 export type EventHandler<D> = Array<EventHandlerObject<D>>
 
-export type EventHandlerConfig<D, R, C, A, T> = MaybeArray<
-  ActionConfig<D, A> | EventHandlerObjectConfig<D, R, C, A, T>
+export type EventHandlerDesign<D, R, C, A, T> = MaybeArray<
+  ActionDesign<D, A> | EventHandlerObjectDesign<D, R, C, A, T>
 >
 
 export type RepeatEventHandler<D> = {
@@ -143,9 +143,9 @@ export type RepeatEventHandler<D> = {
   delay?: Time<D>
 }
 
-export type RepeatEventHandlerConfig<D, R, C, A, T> = {
-  event: EventHandlerConfig<D, R, C, A, T>
-  delay?: TimeConfig<D, T>
+export type RepeatEventHandlerDesign<D, R, C, A, T> = {
+  event: EventHandlerDesign<D, R, C, A, T>
+  delay?: TimeDesign<D, T>
 }
 // Async Event Handler
 
@@ -155,10 +155,10 @@ export type AsyncEventHandler<D> = {
   onReject?: EventHandler<D>
 }
 
-export type AsyncEventHandlerConfig<D, R, C, A, Y, T> = {
+export type AsyncEventHandlerDesign<D, R, C, A, Y, T> = {
   await: Extract<keyof Y, string> | Async<D>
-  onResolve: EventHandlerConfig<D, R, C, A, T>
-  onReject?: EventHandlerConfig<D, R, C, A, T>
+  onResolve: EventHandlerDesign<D, R, C, A, T>
+  onReject?: EventHandlerDesign<D, R, C, A, T>
 }
 
 // Verbose Logging Types
@@ -199,20 +199,20 @@ export interface State<D> {
   initial?: string
 }
 
-export interface StateConfig<D, R, C, A, Y, T, V> {
-  on?: Record<string, EventHandlerConfig<D, R, C, A, T>>
-  onEnter?: EventHandlerConfig<D, R, C, A, T>
-  onExit?: EventHandlerConfig<D, R, C, A, T>
-  onEvent?: EventHandlerConfig<D, R, C, A, T>
-  repeat?: RepeatEventHandlerConfig<D, R, C, A, T>
-  async?: AsyncEventHandlerConfig<D, R, C, A, Y, T>
-  states?: Record<string, StateConfig<D, R, C, A, Y, T, V>>
+export interface StateDesign<D, R, C, A, Y, T, V> {
+  on?: Record<string, EventHandlerDesign<D, R, C, A, T>>
+  onEnter?: EventHandlerDesign<D, R, C, A, T>
+  onExit?: EventHandlerDesign<D, R, C, A, T>
+  onEvent?: EventHandlerDesign<D, R, C, A, T>
+  repeat?: RepeatEventHandlerDesign<D, R, C, A, T>
+  async?: AsyncEventHandlerDesign<D, R, C, A, Y, T>
+  states?: Record<string, StateDesign<D, R, C, A, Y, T, V>>
   initial?: string
 }
 
-// Config
+// Design
 
-export interface Config<
+export interface Design<
   D,
   R extends Record<string, Result<D>> = any,
   C extends Record<string, Condition<D>> = any,
@@ -220,7 +220,7 @@ export interface Config<
   Y extends Record<string, Async<D>> = any,
   T extends Record<string, Time<D>> = any,
   V extends Record<string, Value<D>> = any
-> extends StateConfig<D, R, C, A, Y, T, V> {
+> extends StateDesign<D, R, C, A, Y, T, V> {
   id?: string
   data?: D
   results?: R
@@ -231,7 +231,7 @@ export interface Config<
   values?: V
 }
 
-export interface ConfigWithHelpers<
+export interface DesignWithHelpers<
   D,
   R extends Record<string, Result<D>> = any,
   C extends Record<string, Condition<D>> = any,
@@ -239,29 +239,29 @@ export interface ConfigWithHelpers<
   Y extends Record<string, Async<D>> = any,
   T extends Record<string, Time<D>> = any,
   V extends Record<string, Value<D>> = any
-> extends Config<D, R, C, A, Y, T, V> {
-  createEventHandlerConfig: (
-    config: EventHandlerConfig<D, R, C, A, T>
-  ) => EventHandlerConfig<D, R, C, A, T>
-  createEventHandlerObjectConfig: (
-    config: EventHandlerObjectConfig<D, R, C, A, T>
-  ) => EventHandlerObjectConfig<D, R, C, A, T>
-  createAsyncEventHandlerConfig: (
-    config: AsyncEventHandlerConfig<D, R, C, A, Y, T>
-  ) => AsyncEventHandlerConfig<D, R, C, A, Y, T>
-  createRepeatEventHandlerConfig: (
-    config: RepeatEventHandlerConfig<D, R, C, A, T>
-  ) => RepeatEventHandlerConfig<D, R, C, A, T>
-  createStateConfig: (
-    config: StateConfig<D, R, C, A, Y, T, V>
-  ) => StateConfig<D, R, C, A, Y, T, V>
-  createResultConfig: (config: ResultConfig<D, R>) => ResultConfig<D, R>
-  createConditionConfig: (
-    config: ConditionConfig<D, C>
-  ) => ConditionConfig<D, C>
-  createActionConfig: (config: ActionConfig<D, A>) => ActionConfig<D, A>
-  createTimeConfig: (config: TimeConfig<D, T>) => TimeConfig<D, T>
-  createValueConfig: (config: Value<D>) => Value<D>
+> extends Design<D, R, C, A, Y, T, V> {
+  createEventHandlerDesign: (
+    config: EventHandlerDesign<D, R, C, A, T>
+  ) => EventHandlerDesign<D, R, C, A, T>
+  createEventHandlerObjectDesign: (
+    config: EventHandlerObjectDesign<D, R, C, A, T>
+  ) => EventHandlerObjectDesign<D, R, C, A, T>
+  createAsyncEventHandlerDesign: (
+    config: AsyncEventHandlerDesign<D, R, C, A, Y, T>
+  ) => AsyncEventHandlerDesign<D, R, C, A, Y, T>
+  createRepeatEventHandlerDesign: (
+    config: RepeatEventHandlerDesign<D, R, C, A, T>
+  ) => RepeatEventHandlerDesign<D, R, C, A, T>
+  createState: (
+    config: StateDesign<D, R, C, A, Y, T, V>
+  ) => StateDesign<D, R, C, A, Y, T, V>
+  createResultDesign: (config: ResultDesign<D, R>) => ResultDesign<D, R>
+  createConditionDesign: (
+    config: ConditionDesign<D, C>
+  ) => ConditionDesign<D, C>
+  createActionDesign: (config: ActionDesign<D, A>) => ActionDesign<D, A>
+  createTimeDesign: (config: TimeDesign<D, T>) => TimeDesign<D, T>
+  createValueDesign: (config: Value<D>) => Value<D>
 }
 
 // Subscribers
@@ -274,11 +274,11 @@ export type SubscriberFn<
   Y extends Record<string, Async<D>> = any,
   T extends Record<string, Time<D>> = any,
   V extends Record<string, Value<D>> = any
-> = (update: StateDesign<D, R, C, A, Y, T, V>) => void
+> = (update: DesignedState<D, R, C, A, Y, T, V>) => void
 
 // State Design
 
-export interface StateDesign<
+export interface DesignedState<
   D,
   R extends Record<string, Result<D>>,
   C extends Record<string, Condition<D>>,
@@ -296,7 +296,7 @@ export interface StateDesign<
   send: (
     eventName: string,
     payload?: any
-  ) => Promise<StateDesign<D, R, C, A, Y, T, V>>
+  ) => Promise<DesignedState<D, R, C, A, Y, T, V>>
   can: (eventName: string, payload?: any) => boolean
   isIn: (...paths: string[]) => boolean
   isInAny: (...paths: string[]) => boolean
@@ -310,7 +310,7 @@ export interface StateDesign<
     ) => any,
     initial?: any
   ) => any
-  getConfig: () => Config<D, R, C, A, Y, T, V>
-  clone: () => StateDesign<D, R, C, A, Y, T, V>
+  getDesign: () => Design<D, R, C, A, Y, T, V>
+  clone: () => DesignedState<D, R, C, A, Y, T, V>
   values: Values<D, V>
 }