+++ packages/state-designer/test/imports.test.tsx
@@ -1,44 +1,22 @@
-import * as React from "react"
-import * as ReactDOM from "react-dom"
+import { createDesign, createState } from "../src"
 
-import { createConfig, createStateDesign, useStateDesign } from "../src"
-
-const config = createConfig({
+const config = createDesign({
   data: { count: 0 },
   on: {
     ADDED_ONE: (d) => d.count++,
   },
 })
 
-const state = createStateDesign(config)
+const state = createState(config)
 
-describe("createConfig", () => {
+describe("createDesign", () => {
   it("Should create a config.", () => {
     expect(config).toBeTruthy()
   })
 })
 
-describe("createStateDesign", () => {
+describe("createState", () => {
   it("Should create a state.", () => {
     expect(state).toBeTruthy()
-  })
-})
-
-describe("useStateDesign", () => {
-  it("should support the hook.", () => {
-    const App = () => {
-      const { data, send } = useStateDesign(state)
-
-      return (
-        <div>
-          <h1>{data.count}</h1>
-          <button onClick={() => send("ADDED_ONE")}>+1</button>
-        </div>
-      )
-    }
-
-    const div = document.createElement("div")
-    ReactDOM.render(<App />, div)
-    ReactDOM.unmountComponentAtNode(div)
   })
 })