+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.2.0
+
+- Changes core API
+  - `createStateDesigner` changed to `createState`
+  - `createConfig` changed to `createDesign`
+  - Changed `state.getConfig` to `state.getDesign`.
+  - Keeps `useStateDesigner`.
+- Changes types.
+- Expands docs. Terminology here is complex, so let's follow it in the types and API too.
+
 ## 1.1.22
 
 - Restores `default` option in `whenIn`. If none of the other provided state paths are active, the `default` option will be used instead.
@@ -12,12 +22,12 @@
 ## 1.1.20
 
 - Removes `elseTo`, `elseDo`, and `elseSend`. Introduces `else`, which takes an event handler. This allows event handlers to nest infinitely, if you're intro that kind of thing.
-- Adds debugging callback as second argument of `createStateDesign`. The callback will receive a LOT of messages, but you can switch on the callback's second argument (a string like "async events" or "actions") to only log out the messages that matter to you.
+- Adds debugging callback as second argument of `createState`. The callback will receive a LOT of messages, but you can switch on the callback's second argument (a string like "async events" or "actions") to only log out the messages that matter to you.
 - Adds the `secretlyDo` option. Run an action without causing the state to update. (Normally, running a `do` action or transition will trigger an update.) Useful when working with animation libraries in React, such as Framer Motion, that use alternative data flow.
 
 ```jsx
 const mouseX = React.useRef(0)
-const { send } = useStateDesign({
+const { send } = useStateDesigner({
   on: {
     MOUSE_MOVED: {
       secretlyDo: (data, payload) => (mouseX.current = payload),
@@ -57,7 +67,7 @@ const { send } = useStateDesign({
 
 ## 1.1.5
 
-- Adds helpers to the config returned from `createConfig`. This is useful (in theory) when composing a configuration from separate parts, rather than in one giant object, and where type safety is required.
+- Adds helpers to the config returned from `createDesign`. This is useful (in theory) when composing a configuration from separate parts, rather than in one giant object, and where type safety is required.
 
 ## 1.1.4
 
@@ -74,8 +84,8 @@ const { send } = useStateDesign({
 ## 1.1.0
 
 - Rewrote entire library, created monorepo to split core from react.
-- @state-designer/core includes only the `createStateDesign` and `createConfig` functions, and all types as `S`.
-- @state-designer/react includes only the `useStateDesign` hook.
+- @state-designer/core includes only the `createState` and `createDesign` functions, and all types as `S`.
+- @state-designer/react includes only the `useStateDesigner` hook.
 - The same APIs all work, but the underlying code is entirely rewritten and new.
 - Moved from classes to functions.
 - Logic is improved, bugs fixed.